Rename dir for zod schema in acl request
