+++ coral/src/app/features/topics/acl-request/form-schemas/topic-acl-request-consumer.ts
@@ -6,11 +6,11 @@ import {
   remarks,
   topicname,
   teamId,
-} from "src/app/features/topics/acl-request/schemas/topic-acl-request-shared-fields";
+} from "src/app/features/topics/acl-request/form-schemas/topic-acl-request-shared-fields";
 import {
   hasOnlyValidCharacters,
   validateAclPrincipleValue,
-} from "src/app/features/topics/acl-request/schemas/validation";
+} from "src/app/features/topics/acl-request/form-schemas/validation";
 import { z } from "zod";
 
 const consumergroup = z
