+++ coral/src/app/features/topics/acl-request/forms/TopicConsumerForm.tsx
@@ -21,7 +21,7 @@ import EnvironmentField from "src/app/features/topics/acl-request/fields/Environ
 import IpOrPrincipalField from "src/app/features/topics/acl-request/fields/IpOrPrincipalField";
 import RemarksField from "src/app/features/topics/acl-request/fields/RemarksField";
 import TopicNameField from "src/app/features/topics/acl-request/fields/TopicNameField";
-import { TopicConsumerFormSchema } from "src/app/features/topics/acl-request/schemas/topic-acl-request-consumer";
+import { TopicConsumerFormSchema } from "src/app/features/topics/acl-request/form-schemas/topic-acl-request-consumer";
 import { createAclRequest } from "src/domain/acl/acl-api";
 import { Environment } from "src/domain/environment";
 import { parseErrorMsg } from "src/services/mutation-utils";
